Split-off job-offer and -search useCases
