+++ webofneeds/won-owner-webapp/src/main/webapp/config/usecases/professional/job-search.js
@@ -1,0 +1,42 @@
+import { details, emptyDraft } from "../../detail-definitions.js";
+import { interestsDetail, skillsDetail } from "../../details/person.js";
+import { findLatestIntervallEndInJsonLdOrNowAndAddMillis } from "../../../app/won-utils.js";
+import {
+  industryDetail,
+  employmentTypesDetail,
+  organizationNamesDetail,
+} from "../../details/jobs.js";
+import { jobLocation } from "../../details/location.js";
+
+export const jobSearch = {
+  identifier: "jobSearch",
+  label: "Search a Job",
+  // icon: "#ico36_uc_find_people", TODO
+  doNotMatchAfter: findLatestIntervallEndInJsonLdOrNowAndAddMillis,
+  draft: {
+    ...emptyDraft,
+    is: {
+      "@type": "s:Person",
+      tags: ["search-job"],
+    },
+    seeks: {
+      "@type": "s:JobPosting",
+    },
+    searchString: ["offer-job"],
+  },
+  isDetails: {
+    title: { ...details.title },
+    description: { ...details.description },
+    // location: { ...details.location }, // why would your current location of residency matter?
+    // person: { ...details.person }, // has current company fields, that are weird for a search
+    skills: { ...skillsDetail },
+    interests: { ...interestsDetail },
+  },
+  seeksDetails: {
+    description: { ...details.description },
+    jobLocation: { ...jobLocation },
+    industry: { ...industryDetail },
+    employmentTypes: { ...employmentTypesDetail },
+    organizationNames: { ...organizationNamesDetail },
+  },
+};
